+++ example/a.ts
@@ -1,21 +1,27 @@
 import { Elysia, t } from '../src'
 import { req } from '../test/utils'
 
-const app = new Elysia()
-	.model({
-		string: t.String()
-	})
-	.get('/', () => ({ a: 'a' }), {
-		response: t.Object({
-			a: t.Ref('string')
+const app = new Elysia().get(
+	'/test',
+	({ query: { id } }) => ({
+		id,
+		type: typeof id
+	}),
+	{
+		query: t.Object({
+			id: t
+				.Transform(t.Array(t.UnionEnum(['test', 'foo'])))
+				.Decode((id) => ({ value: id }))
+				.Encode((id) => id.value)
 		})
-	})
-	.listen(3000)
+	}
+)
 
-app.handle(req('/'))
-	.then((x) => x.json())
-	.then(console.log)
-
-console.log(
-	`🦊 Elysia is running at ${app.server?.hostname}:${app.server?.port}`
+app.handle(req('/test?id=test'))
+	.then((x) =>
+	x.json().then((v) =>
+		console.dir(v, {
+			depth: null
+		})
+	)
 )