+++ example/file.ts
@@ -1,15 +1,19 @@
-import { Elysia } from '../src'
+import { Elysia, file } from '../src'
+import Page from './index.html'
 
 /**
  * Example of handle single static file
  *
- * For more advavnce use-case, eg. folder
  * @see https://github.com/elysiajs/elysia-static
  */
 new Elysia()
-	.get('/tako', ({ set }) => {
-		set.headers.server = 'Elysia'
-
-		return Bun.file('./example/takodachi.png')
+	.headers({
+		server: 'Elysia'
 	})
+	.onRequest(({ request }) => {
+		console.log(request.method, request.url)
+	})
+	.get('/', Page)
+	.get('/tako', file('./example/takodachi.png'))
+	.get('/mika.mp4', Bun.file('test/kyuukurarin.mp4'))
 	.listen(3000)