+++ src/compose.ts
@@ -45,7 +45,7 @@ import type {
 	LifeCycleStore,
 	SchemaValidator
 } from './types'
-import type { TypeCheck } from './type-system'
+import { type TypeCheck } from './type-system'
 
 const TypeBoxSymbol = {
 	optional: Symbol.for('TypeBox.Optional'),
@@ -229,11 +229,13 @@ const createReport = ({
 const composeValidationFactory = ({
 	injectResponse = '',
 	normalize = false,
-	validator
+	validator,
+	encodeSchema = false
 }: {
 	injectResponse?: string
 	normalize?: boolean
 	validator: SchemaValidator
+	encodeSchema?: boolean
 }) => ({
 	composeValidation: (type: string, value = `c.${type}`) =>
 		`c.set.status=422;throw new ValidationError('${type}',validator.${type},${value})`,
@@ -266,8 +268,17 @@ const composeValidationFactory = ({
 				`throw new ValidationError('response',validator.response['${status}'],${name})` +
 				'}' +
 				`c.set.status = ${status}` +
-				'}' +
-				'break\n'
+				'}\n'
+
+			if (
+				encodeSchema &&
+				// @ts-expect-error hasTransform is appended by getResponseSchemaValidator
+				(value.hasTransform || typeof value.Decode === 'function')
+			) {
+				code += `${name}=validator.response['${status}'].Encode(${name})\n`
+			}
+
+			code += 'break\n'
 		}
 
 		return code + '}'
@@ -646,11 +657,17 @@ export const composeHandler = ({
 
 	const hasQuery = inference.query || !!validator.query
 
+	const requestNoBody =
+		hooks.parse?.length === 1 &&
+		// @ts-expect-error
+		hooks.parse[0].fn === 'none'
+
 	const hasBody =
 		method !== '$INTERNALWS' &&
 		method !== 'GET' &&
 		method !== 'HEAD' &&
-		(inference.body || !!validator.body || !!hooks.parse?.length)
+		(inference.body || !!validator.body || !!hooks.parse?.length) &&
+		!requestNoBody
 
 	if (hasBody) fnLiteral += `let isParsing=false\n`
 
@@ -720,11 +737,13 @@ export const composeHandler = ({
 	}
 
 	const normalize = app.config.normalize
+	const encodeSchema = app.config.experimental?.encodeSchema
 
 	const { composeValidation, composeResponseValidation } =
 		composeValidationFactory({
 			normalize,
-			validator
+			validator,
+			encodeSchema
 		})
 
 	if (hasHeaders) fnLiteral += adapter.headers
@@ -869,7 +888,7 @@ export const composeHandler = ({
 				'}'
 		} else {
 			fnLiteral +=
-				'if(c.qi!==-1){' + `let url = '&' + c.url.slice(c.qi + 1)\n`
+				'if(c.qi!==-1){' + `let url='&'+c.url.slice(c.qi + 1)\n`
 
 			let index = 0
 			for (const {