+++ test/core/modules.test.ts
@@ -209,7 +209,7 @@ describe('Modules', () => {
 		expect(response.status).toBe(200)
 	})
 
-	it('recompile async plugin once registered', async () => {
+	it('recompile nested async plugin once registered', async () => {
 		const asyncPlugin = Promise.resolve(new Elysia({ name: 'AsyncPlugin' }))
 
 		const plugin = new Elysia({ name: 'Plugin' })
@@ -228,5 +228,24 @@ describe('Modules', () => {
 		// If the plugin doesn't recompile, route index
 		// would be pointed to /foo instead of /plugin
 		expect(response).toEqual('GET /plugin')
+	})
+
+	it('recompile not nested async plugin once registered', async () => {
+		const asyncPlugin = Promise.resolve(new Elysia({ name: 'AsyncPlugin' }))
+
+		const plugin = new Elysia({ name: 'Plugin' })
+			.use(asyncPlugin)
+			.get('/plugin', () => 'GET /plugin')
+
+		const app = new Elysia({ name: 'App' })
+			.use(plugin)
+			.get('/foo', () => 'GET /foo')
+
+		const response = await app.handle(
+			new Request('http://localhost/plugin')
+		)
+
+		const text = await response.text()
+		expect(text).toEqual('GET /plugin')
 	})
 })