+++ test/type-system/array-string.test.ts
@@ -6,7 +6,17 @@ import { req } from '../utils'
 
 describe('TypeSystem - ArrayString', () => {
 	it('Create', () => {
-		expect(Value.Create(t.ArrayString())).toBe('[]')
+		// @ts-expect-error
+		expect(Value.Create(t.ArrayString())).toBe(undefined)
+
+		expect(
+			Value.Create(
+				t.ArrayString(t.Any(), {
+					default: '[]'
+				})
+			)
+			// @ts-expect-error
+		).toBe('[]')
 	})
 
 	it('Check', () => {
@@ -18,63 +28,55 @@ describe('TypeSystem - ArrayString', () => {
 	it('Encode', () => {
 		const schema = t.ArrayString(t.Number())
 
-		expect(
-			Value.Encode<typeof schema, string>(schema, [1])
-		).toBe(JSON.stringify([1]))
+		expect(Value.Encode<typeof schema, string>(schema, [1])).toBe(
+			JSON.stringify([1])
+		)
 
-		expect(
-			Value.Encode<typeof schema, string>(schema, [1])
-		).toBe(JSON.stringify([1]))
+		expect(Value.Encode<typeof schema, string>(schema, [1])).toBe(
+			JSON.stringify([1])
+		)
 	})
 
 	it('Decode', () => {
 		const schema = t.ArrayString(t.Number())
 
-		expect(
-			Value.Decode<typeof schema>(
-				schema,
-				'[1]'
-			)
-		).toEqual([1])
+		expect(Value.Decode<typeof schema>(schema, '[1]')).toEqual([1])
 
-		expect(() =>
-			Value.Decode<typeof schema>(
-				schema,
-				'1'
-			)
-		).toThrow()
+		expect(() => Value.Decode<typeof schema>(schema, '1')).toThrow()
 	})
 
 	it('Integrate', async () => {
-		const app = new Elysia().post(
-			'/',
-			({ body }) => body,
-			{
-				body: t.Object({
-					id: t.ArrayString(t.Number())
-				})
-			}
-		)
+		const app = new Elysia().post('/', ({ body }) => body, {
+			body: t.Object({
+				id: t.ArrayString(t.Number())
+			})
+		})
 
-		const res1 = await app.handle(new Request('http://localhost', {
-			method: "POST",
-			headers: { "Content-Type": "application/json" },
-			body: JSON.stringify({ id: JSON.stringify([1, 2, 3]) })
-		}))
+		const res1 = await app.handle(
+			new Request('http://localhost', {
+				method: 'POST',
+				headers: { 'Content-Type': 'application/json' },
+				body: JSON.stringify({ id: JSON.stringify([1, 2, 3]) })
+			})
+		)
 		expect(res1.status).toBe(200)
 
-		const res2 = await app.handle(new Request('http://localhost', {
-			method: "POST",
-			headers: { "Content-Type": "application/json" },
-			body: JSON.stringify({ id: [1, 2, 3] })
-		}))
+		const res2 = await app.handle(
+			new Request('http://localhost', {
+				method: 'POST',
+				headers: { 'Content-Type': 'application/json' },
+				body: JSON.stringify({ id: [1, 2, 3] })
+			})
+		)
 		expect(res2.status).toBe(200)
 
-		const res3 = await app.handle(new Request('http://localhost', {
-			method: "POST",
-			headers: { "Content-Type": "application/json" },
-			body: JSON.stringify({ id: ['a', 2, 3] })
-		}))
+		const res3 = await app.handle(
+			new Request('http://localhost', {
+				method: 'POST',
+				headers: { 'Content-Type': 'application/json' },
+				body: JSON.stringify({ id: ['a', 2, 3] })
+			})
+		)
 		expect(res3.status).toBe(422)
 	})
 })